requirements.txt
cmake==3.17.2
dlib==19.18.0
opencv-python==4.6.0.66
Pillow~=10.3.0
face_recognition==1.3.0


mediapipe~=0.10.14
numpy~=1.26.4
torch~=2.2.0
tqdm~=4.66.4

deepface~=0.0.92
tf_keras
streamlit~=1.36.0
gradio~=3.50.2
pymongo~=4.8.0